refactoring: move status command to a new component teambit.component/status (#5596)
