+++ .bitmap
@@ -875,6 +875,13 @@
         "mainFile": "index.ts",
         "rootDir": "scopes/component/snapping"
     },
+    "status": {
+      "scope": "",
+      "version": "",
+      "defaultScope": "teambit.component",
+      "mainFile": "index.ts",
+      "rootDir": "scopes/component/status"
+  },
     "string/capitalize": {
         "scope": "teambit.toolbox",
         "version": "0.0.483",
