+++ e2e/commands/add.e2e.1.ts
@@ -5,8 +5,7 @@ import * as path from 'path';
 
 import { InvalidName } from '@teambit/legacy-bit-id';
 import AddTestsWithoutId from '../../src/cli/commands/exceptions/add-tests-without-id';
-import { statusInvalidComponentsMsg } from '../../src/cli/commands/public-cmds/status-cmd';
-import { AUTO_GENERATED_MSG } from '../../src/constants';
+import { AUTO_GENERATED_MSG, statusInvalidComponentsMsg } from '../../src/constants';
 import { MissingMainFile } from '../../src/consumer/bit-map/exceptions';
 import {
   ExcludedMainFile,
