+++ e2e/commands/status.e2e.2.ts
@@ -2,9 +2,8 @@ import chai, { expect } from 'chai';
 import fs from 'fs-extra';
 import * as path from 'path';
 import { IssuesClasses, MISSING_DEPS_SPACE, MISSING_NESTED_DEPS_SPACE } from '@teambit/component-issues';
-import { statusFailureMsg, statusInvalidComponentsMsg } from '../../src/cli/commands/public-cmds/status-cmd';
 import { MISSING_PACKAGES_FROM_OVERRIDES_LABEL } from '../../src/cli/templates/component-issues-template';
-import { IMPORT_PENDING_MSG } from '../../src/constants';
+import { IMPORT_PENDING_MSG, statusFailureMsg, statusInvalidComponentsMsg } from '../../src/constants';
 import ComponentNotFoundInPath from '../../src/consumer/component/exceptions/component-not-found-in-path';
 import MissingFilesFromComponent from '../../src/consumer/component/exceptions/missing-files-from-component';
 import Helper from '../../src/e2e-helper/e2e-helper';
