+++ e2e/flows/dist-outside-components.e2e.2.ts
@@ -1,8 +1,8 @@
 import chai, { expect } from 'chai';
 import fs from 'fs-extra';
 import * as path from 'path';
+import { statusFailureMsg } from '../../src/constants';
 
-import { statusFailureMsg } from '../../src/cli/commands/public-cmds/status-cmd';
 import Helper from '../../src/e2e-helper/e2e-helper';
 import * as fixtures from '../../src/fixtures/fixtures';
