+++ scopes/component/status/index.ts
@@ -1,0 +1,5 @@
+import { StatusAspect } from './status.aspect';
+
+export type { StatusMain } from './status.main.runtime';
+export default StatusAspect;
+export { StatusAspect };
