+++ scopes/component/status/status.aspect.ts
@@ -1,0 +1,5 @@
+import { Aspect } from '@teambit/harmony';
+
+export const StatusAspect = Aspect.create({
+  id: 'teambit.component/status',
+});
