+++ scopes/harmony/bit/manifests.ts
@@ -79,6 +79,7 @@ import { MochaAspect } from '@teambit/mocha';
 import { BitCustomAspectAspect } from '@teambit/bit-custom-aspect';
 import { CommunityAspect } from '@teambit/community';
 import { CloudAspect } from '@teambit/cloud';
+import { StatusAspect } from '@teambit/status';
 import { SnappingAspect } from '@teambit/snapping';
 import { MergingAspect } from '@teambit/merging';
 import { BitAspect } from './bit.aspect';
@@ -162,6 +163,7 @@ export const manifestsMap = {
   [MochaAspect.id]: MochaAspect,
   [BitCustomAspectAspect.id]: BitCustomAspectAspect,
   [DiagnosticAspect.id]: DiagnosticAspect,
+  [StatusAspect.id]: StatusAspect,
   [CommunityAspect.id]: CommunityAspect,
   [CloudAspect.id]: CloudAspect,
   [SnappingAspect.id]: SnappingAspect,
