+++ src/cli/command-registry-builder.ts
@@ -44,7 +44,6 @@ import Remote from './commands/public-cmds/remote-cmd';
 import Remove from './commands/public-cmds/remove-cmd';
 import ScopeConfig from './commands/public-cmds/scope-config-cmd';
 import Show from './commands/public-cmds/show-cmd';
-import Status from './commands/public-cmds/status-cmd';
 import Test from './commands/public-cmds/test-cmd';
 import Untag from './commands/public-cmds/untag-cmd';
 import Untrack from './commands/public-cmds/untrack-cmd';
@@ -64,8 +63,6 @@ export default function registerCommands(extensionsCommands: Array<Commands>): C
       new Isolate(),
       // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
       new Import(),
-      // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
-      new Status(),
       // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
       new List(),
       new Config(),
